+++ tests/integration/test_run_download.py
@@ -84,6 +84,7 @@ logging:
 
 runtime:
   n_download_threads: 2
+  n_cutout_processes: 2
   bands: ["whigs-g", "cfis_lsb-r", "ps-i"]
   resume: false
 
@@ -92,6 +93,12 @@ tiles:
   show_tile_statistics: false
   band_constraint: 1
   require_all_specified_bands: false
+
+cutouts:
+    enable: true
+    cutouts_only: false
+    size_pix: 256
+    output_subdir: "cutouts"
 
 inputs:
   source: "tiles"